feat(prog1): actividad prog1 con README en carpeta 'Arrays' con 3 ejercicios
